Fixed getDimension return type.

+++ aikido/include/aikido/statespace/SE2StateSpace.hpp
@@ -100,7 +100,7 @@ public:
                   StateSpace::State *_out) const override;
 
   // Documentation inherited
-  unsigned int getDimension() const override;
+  size_t getDimension() const override;
 
   // Documentation inherited
   void copyState(StateSpace::State* _destination,
